still tracing problem in compute shader
