+++ shaders/update-sprites.wgsl
@@ -4,7 +4,7 @@ struct Particle {
 }
 
 struct SimParams {
-  deltaT: f32,
+  delta_t: f32,
   rule1_distance: f32,
   rule2_distance: f32,
   rule3_distance: f32,
@@ -28,6 +28,7 @@ fn main(@builtin(global_invocation_id) GlobalInvocationID: vec3<u32>) {
 
   var v_pos = particles_a.particles[index].pos;
   var v_vel = particles_a.particles[index].vel;
+
   var c_mass = vec3(0.0);
   var c_vel = vec3(0.0);
   var col_vel = vec3(0.0);
@@ -37,9 +38,9 @@ fn main(@builtin(global_invocation_id) GlobalInvocationID: vec3<u32>) {
   var vel: vec3<f32>;
 
   for (var i = 0u; i < arrayLength(&particles_a.particles); i++) {
-    if (i == index) {
-      continue;
-    }
+    // if (i == index) {
+    //   continue;
+    // }
 
     pos = vec3<f32>(particles_a.particles[i].pos.xy, 0.);
     vel = vec3<f32>(particles_a.particles[i].vel.xy, 0.);
@@ -62,29 +63,33 @@ fn main(@builtin(global_invocation_id) GlobalInvocationID: vec3<u32>) {
     c_vel /= f32(c_vel_count);
   }
   v_vel += (c_mass * params.rule1_scale) + (col_vel * params.rule2_scale) + (c_vel * params.rule3_scale);
+  v_vel = vec3<f32>(0.05,0.0,0.0);
 
   // clamp velocity for a more pleasing simulation
-  v_vel = normalize(v_vel) * clamp(length(v_vel), 0.0, 0.1);
+  // v_vel = normalize(v_vel) * clamp(length(v_vel), 0.0, 0.1);
   // kinematic update
-  v_pos = v_pos + (v_vel * params.deltaT);
+  v_pos = v_pos + (v_vel * params.delta_t);
   // Wrap around boundary
-  if (v_pos.x < -1.0) {
-    v_pos.x = 1.0;
-  }
-  if (v_pos.x > 1.0) {
-    v_pos.x = -1.0;
-  }
-  if (v_pos.y < -1.0) {
-    v_pos.y = 1.0;
-  }
-  if (v_pos.y > 1.0) {
-    v_pos.y = -1.0;
-  }
+  // if (v_pos.x < -1.0) {
+  //   v_pos.x = 1.0;
+  // }
+  // if (v_pos.x > 1.0) {
+  //   v_pos.x = -1.0;
+  // }
+  // if (v_pos.y < -1.0) {
+  //   v_pos.y = 1.0;
+  // }
+  // if (v_pos.y > 1.0) {
+  //   v_pos.y = -1.0;
+  // }
   // Write back
-  particles_b.particles[index].pos = v_pos;
-  particles_b.particles[index].vel = v_vel;
-  // particles_b.particles[index].pos = particles_b.particles[index].pos;
-  // particles_b.particles[index].vel = particles_b.particles[index].vel;
+  // particles_b.particles[index].pos = vec3(0.3,0.0,0.0);
+  // particles_b.particles[index].vel = v_vel;
+  particles_b.particles[index].pos = particles_b.particles[index].pos;
+  particles_b.particles[index].vel = particles_b.particles[index].vel;
+
+  particles_b.particles[index].pos = vec3(0.5, 0.1, 0.1);
+  particles_b.particles[index].vel = vec3(0.0, 0.0, 0.0);
 }
 
 fn xy0(v: vec3<f32>) -> vec3<f32> {
